+++ ocbpy/tests/test_pysat.py
@@ -39,7 +39,7 @@ class TestPysatUtils(unittest.TestCase):
         self.ocb = None
         self.added_keys = list()
         self.pysat_keys = list()
-        self.arevectors = list(),
+        self.arevectors = list()
         self.nkeys = 0
 
     def tearDown(self):
@@ -52,103 +52,106 @@ class TestPysatUtils(unittest.TestCase):
         """ Test that metadata was added properly
         """
 
+        # Test passes if no MetaData
         if self.meta is None:
-            self.assertTrue(True)
-        else:
-            self.assertTrue(self.ocb_key in self.meta.keys())
-            if self.pysat_key is not None:
-                self.assertTrue(self.pysat_key in self.meta.keys())
+            return
 
-            if self.isvector:
-                # Test the fill value
-                self.assertIsNone(
-                    self.meta[self.ocb_key][self.meta.fill_label])
-            elif self.pysat_key is not None:
-                # Test the elements that are identical
-                for ll in [self.meta.units_label, self.meta.scale_label,
-                           self.meta.min_label, self.meta.max_label,
-                           self.meta.fill_label]:
-                    try:
-                        if np.isnan(self.meta[self.pysat_key][ll]):
-                            self.assertTrue(
-                                np.isnan(self.meta[self.ocb_key][ll]))
-                        else:
-                            self.assertEqual(self.meta[self.ocb_key][ll],
-                                             self.meta[self.pysat_key][ll])
-                    except TypeError:
-                        if len(self.meta[self.ocb_key][ll]) == 0:
-                            self.assertEqual(
-                                len(self.meta[self.pysat_key][ll]), 0)
-                        else:
-                            self.assertRegex(self.meta[self.ocb_key][ll],
-                                             self.meta[self.pysat_key][ll])
+        # Tests for MetaData
+        self.assertIn(self.ocb_key, list(self.meta.keys()))
+        if self.pysat_key is not None:
+            self.assertIn(self.pysat_key, list(self.meta.keys()))
 
-            # Test the elements that have "OCB" appended to the text
-            for ll in [self.meta.name_label, self.meta.axis_label,
-                       self.meta.plot_label]:
-                sline = self.meta[self.ocb_key][ll].split(" ")
-                self.assertRegex(sline[0], "OCB")
-                if not self.isvector and self.pysat_key is not None:
-                    self.assertRegex(" ".join(sline[1:]),
-                                     self.meta[self.pysat_key][ll])
+        if self.isvector:
+            # Test the fill value
+            self.assertIsNone(
+                self.meta[self.ocb_key][self.meta.labels.fill_val])
+        elif self.pysat_key is not None:
+            # Test the elements that are identical
+            for ll in [self.meta.labels.units, self.meta.labels.min_val,
+                       self.meta.labels.max_val, self.meta.labels.fill_val]:
+                try:
+                    if np.isnan(self.meta[self.pysat_key][ll]):
+                        self.assertTrue(np.isnan(self.meta[self.ocb_key][ll]))
+                    else:
+                        self.assertEqual(self.meta[self.ocb_key][ll],
+                                         self.meta[self.pysat_key][ll])
+                except TypeError:
+                    if len(self.meta[self.ocb_key][ll]) == 0:
+                        self.assertEqual(len(self.meta[self.pysat_key][ll]), 0)
+                    else:
+                        self.assertRegex(self.meta[self.ocb_key][ll],
+                                         self.meta[self.pysat_key][ll])
 
-            # Test the remaining elements
-            self.assertEqual(
-                self.meta[self.ocb_key][self.meta.desc_label].find(
-                    "Open Closed"), 0)
-            if self.notes is not None:
-                self.assertRegex(
-                    self.meta[self.ocb_key][self.meta.notes_label], self.notes)
+        # Test the elements that have "OCB" appended to the text
+        sline = self.meta[self.ocb_key][self.meta.labels.name].split(" ")
+        self.assertRegex(sline[0], "OCB")
+        if not self.isvector and self.pysat_key is not None:
+            self.assertRegex(" ".join(sline[1:]),
+                             self.meta[self.pysat_key][self.meta.labels.name])
 
-            del ll, sline
+        # Test the remaining elements
+        self.assertEqual(self.meta[self.ocb_key][self.meta.labels.desc].find(
+            "Open Closed"), 0)
+        if self.notes is not None:
+            self.assertRegex(self.meta[self.ocb_key][self.meta.labels.notes],
+                             self.notes)
+        return
 
     def test_ocb_added(self):
         """ Test if OCB data was added correctly
         """
+        # If there is no test instrument, test passes
         if self.test_inst is None:
-            self.assertTrue(True)
-        else:
-            if len(self.arevectors) < len(self.added_keys):
-                self.arevectors = [False for okey in self.added_keys]
+            return
 
-            self.assertEqual(len(self.added_keys), self.nkeys)
+        # For a test instrument, evaluate attributes
+        if len(self.arevectors) < len(self.added_keys):
+            self.arevectors = [False for okey in self.added_keys]
 
-            for i, okey in enumerate(self.added_keys):
-                # Test to see that data was added
-                self.assertTrue(okey in self.test_inst.data.columns)
+        self.assertEqual(len(self.added_keys), self.nkeys)
 
-                # Test the metadata
-                self.meta = self.test_inst.meta
-                self.pysat_key = self.pysat_keys[i]
-                self.isvector = self.arevectors[i]
-                self.ocb_key = okey
-                self.test_ocb_metadata()
+        for i, okey in enumerate(self.added_keys):
+            # Test to see that data was added
+            self.assertIn(okey, self.test_inst.variables)
 
-                # Test to see that data within 10 minutes of the test OCBs has
-                # OCB locations and other data is NaN
-                match_data = self.test_inst[okey]
-                if self.arevectors[i]:
-                    mask_data = np.not_equal(match_data, None)
-                else:
-                    mask_data = np.isfinite(match_data)
-                match_data = match_data[mask_data]
+            # Test the metadata
+            self.meta = self.test_inst.meta
+            self.pysat_key = self.pysat_keys[i]
+            self.isvector = self.arevectors[i]
+            self.ocb_key = okey
+            self.test_ocb_metadata()
 
-                self.assertEqual(len(match_data), 464)
-                for ii in match_data.index:
-                    check_time = abs(ii - self.ocb.dtime).min().total_seconds()
-                    self.assertLessEqual(check_time, 600.0)
-                if self.arevectors[i]:
-                    self.assertTrue(isinstance(match_data[0],
-                                               ocbpy.ocb_scaling.VectorData))
-                elif self.pysat_keys[i] is not None:
-                    pysat_data = self.test_inst[self.pysat_keys[i]][mask_data]
-                    rscale = (self.ocb.r / (90.0 - self.ocb.boundary_lat))**2
-                    self.assertGreaterEqual(match_data.min(),
-                                            pysat_data.min() * rscale.min())
-                    self.assertGreaterEqual(pysat_data.max() * rscale.max(),
-                                            match_data.max())
+            # Test to see that data within 10 minutes of the test OCBs has
+            # OCB locations and other data is NaN
+            match_data = self.test_inst[okey]
+            if self.arevectors[i]:
+                mask_data = np.not_equal(match_data, None)
+            else:
+                mask_data = np.isfinite(match_data)
+            match_data = match_data[mask_data]
 
-            del okey, match_data, mask_data, ii, check_time
+            self.assertGreater(len(match_data), 0)
+
+            if hasattr(match_data, "index"):
+                match_time = match_data.index
+            else:
+                match_time = pds.to_datetime(match_data['time'].values)
+
+            for ii in match_time:
+                check_time = abs(ii - self.ocb.dtime).min().total_seconds()
+                self.assertLessEqual(check_time, 600.0)
+            if self.arevectors[i]:
+                self.assertTrue(isinstance(match_data[0],
+                                           ocbpy.ocb_scaling.VectorData))
+            elif self.pysat_keys[i] is not None:
+                pysat_data = self.test_inst[self.pysat_keys[i]][mask_data]
+                rscale = (self.ocb.r / (90.0 - self.ocb.boundary_lat))**2
+                self.assertGreaterEqual(match_data.min(),
+                                        pysat_data.min() * rscale.min())
+                self.assertGreaterEqual(pysat_data.max() * rscale.max(),
+                                        match_data.max())
+
+        return
 
 
 @unittest.skipIf(not no_pysat, "pysat installed, cannot test failure")
@@ -179,7 +182,7 @@ class TestPysatStructure(unittest.TestCase):
         pass
 
     def test_add_ocb_to_data_defaults(self):
-        """ test the add_ocb_to_data function defaults"""
+        """ Test the add_ocb_to_data function defaults"""
         defaults = ocb_pysat.add_ocb_to_data.__defaults__
 
         for i in [0, 1, 8]:
@@ -197,7 +200,7 @@ class TestPysatStructure(unittest.TestCase):
         self.assertRegex(defaults[7], 'default')
 
     def test_add_ocb_to_metadata_defaults(self):
-        """ test the add_ocb_to_metadata function defaults"""
+        """ Test the add_ocb_to_metadata function defaults"""
         defaults = ocb_pysat.add_ocb_to_metadata.__defaults__
 
         for i in [0, 2]:
@@ -210,8 +213,7 @@ class TestPysatStructure(unittest.TestCase):
 class TestPysatMethods(unittest.TestCase):
 
     def setUp(self):
-        """ Initialize the OCBoundary object using the test file, as well as
-        the VectorData object
+        """ Initialize the test class
         """
         self.test_file = path.join(path.dirname(ocbpy.__file__), "tests",
                                    "test_data", "test_north_circle")
@@ -222,14 +224,14 @@ class TestPysatMethods(unittest.TestCase):
                                     hemisphere=1)
         self.ocb.rec_ind = 27
 
-        self.test_inst = pysat.Instrument(str('pysat'), str('testing'),
-                                          tag=str('50400'),
+        self.test_inst = pysat.Instrument('pysat', 'testing', num_samples=50400,
                                           clean_level='clean',
                                           update_files=True,
                                           file_date_range=pds.date_range(
                                               self.ocb.dtime[0],
                                               self.ocb.dtime[-1], freq='1D'))
         self.test_inst.load(date=self.ocb.dtime[self.ocb.rec_ind])
+        self.pysat_var2 = 'dummy2'
 
         self.lwarn = u""
         self.lout = u""
@@ -241,15 +243,17 @@ class TestPysatMethods(unittest.TestCase):
         self.utils.setUp()
 
     def tearDown(self):
+        """ Tear down after each test
+        """
         self.utils.tearDown()
-        del self.test_file, self.log_capture, self.test_inst, self.ocb
-        del self.utils, self.lout, self.lwarn, self.ocb_kw
+        del self.test_file, self.log_capture, self.ocb, self.test_inst
+        del self.utils, self.lout, self.lwarn, self.ocb_kw, self.pysat_var2
 
     def test_add_ocb_to_metadata(self):
         """ Test the metadata adding routine
         """
-        ocb_pysat.add_ocb_to_metadata(self.test_inst, "ocb_test", "dummy1",
-                                      notes="test notes")
+        ocb_pysat.add_ocb_to_metadata(self.test_inst, "ocb_test",
+                                      self.utils.pysat_key, notes="test notes")
 
         self.utils.meta = self.test_inst.meta
         self.utils.notes = 'test notes'
@@ -258,8 +262,9 @@ class TestPysatMethods(unittest.TestCase):
     def test_add_ocb_to_metadata_vector(self):
         """ Test the metadata adding routine for vector data
         """
-        ocb_pysat.add_ocb_to_metadata(self.test_inst, "ocb_test", "dummy1",
-                                      notes="test notes", isvector=True)
+        ocb_pysat.add_ocb_to_metadata(self.test_inst, "ocb_test",
+                                      self.utils.pysat_key, notes="test notes",
+                                      isvector=True)
 
         self.utils.meta = self.test_inst.meta
         self.utils.notes = 'test notes'
@@ -270,8 +275,10 @@ class TestPysatMethods(unittest.TestCase):
         """ Test the overwrite block on metadata adding routine
         """
         self.test_add_ocb_to_metadata()
-        ocb_pysat.add_ocb_to_metadata(self.test_inst, "ocb_test", "dummy1",
-                                      notes="test notes two", overwrite=False)
+        ocb_pysat.add_ocb_to_metadata(self.test_inst, "ocb_test",
+                                      self.utils.pysat_key,
+                                      notes="test notes two",
+                                      overwrite=False)
 
         self.lwarn = u"OCB data already has metadata"
         self.lout = self.log_capture.getvalue()
@@ -280,11 +287,13 @@ class TestPysatMethods(unittest.TestCase):
     def test_overwrite_metadata(self):
         """ Test the overwrite permission on metadata adding routine """
         self.test_add_ocb_to_metadata()
-        ocb_pysat.add_ocb_to_metadata(self.test_inst, "ocb_test", "dummy1",
-                                      notes="test notes two", overwrite=True)
+        ocb_pysat.add_ocb_to_metadata(self.test_inst, "ocb_test",
+                                      self.utils.pysat_key,
+                                      notes="test notes two",
+                                      overwrite=True)
 
         meta = self.test_inst.meta
-        self.assertRegex(meta['ocb_test'][meta.notes_label], "test notes two")
+        self.assertRegex(meta['ocb_test'][meta.labels.notes], "test notes two")
 
         del meta
 
@@ -299,7 +308,7 @@ class TestPysatMethods(unittest.TestCase):
         self.utils.nkeys = len(self.utils.added_keys)
         self.utils.pysat_keys = [aa.split("_ocb")[0]
                                  for aa in self.utils.added_keys]
-        self.assertTrue('r_corr' in self.utils.pysat_keys)
+        self.assertIn('r_corr', self.utils.pysat_keys)
         self.utils.pysat_keys[self.utils.pysat_keys.index("r_corr")] = None
 
         self.utils.test_inst = self.test_inst
@@ -318,7 +327,7 @@ class TestPysatMethods(unittest.TestCase):
         self.utils.nkeys = len(self.utils.added_keys)
         self.utils.pysat_keys = [aa.split("_ocb")[0]
                                  for aa in self.utils.added_keys]
-        self.assertTrue('r_corr' in self.utils.pysat_keys)
+        self.assertIn('r_corr', self.utils.pysat_keys)
         self.utils.pysat_keys[self.utils.pysat_keys.index("r_corr")] = None
         self.arevectors = [False for i in range(self.utils.nkeys)]
         self.utils.test_ocb_added()
@@ -327,14 +336,15 @@ class TestPysatMethods(unittest.TestCase):
         """ Test adding ocb to pysat with E-field related variables
         """
         ocb_pysat.add_ocb_to_data(self.test_inst, "latitude", "mlt",
-                                  evar_names=['dummy1'], ocb=self.ocb)
+                                  evar_names=[self.utils.pysat_key],
+                                  ocb=self.ocb)
 
         self.utils.added_keys = [kk for kk in self.test_inst.meta.keys()
                                  if kk.find('_ocb') > 0]
         self.utils.nkeys = len(self.utils.added_keys)
         self.utils.pysat_keys = [aa.split("_ocb")[0]
                                  for aa in self.utils.added_keys]
-        self.assertTrue('r_corr' in self.utils.pysat_keys)
+        self.assertIn('r_corr', self.utils.pysat_keys)
         self.utils.pysat_keys[self.utils.pysat_keys.index("r_corr")] = None
         self.arevectors = [False for i in range(self.utils.nkeys)]
 
@@ -344,14 +354,15 @@ class TestPysatMethods(unittest.TestCase):
         """ Test adding ocb to pysat with Curl E-field related variables
         """
         ocb_pysat.add_ocb_to_data(self.test_inst, "latitude", "mlt",
-                                  curl_evar_names=['dummy2'], ocb=self.ocb)
+                                  curl_evar_names=[self.pysat_var2],
+                                  ocb=self.ocb)
 
         self.utils.added_keys = [kk for kk in self.test_inst.meta.keys()
                                  if kk.find('_ocb') > 0]
         self.utils.nkeys = len(self.utils.added_keys)
         self.utils.pysat_keys = [aa.split("_ocb")[0]
                                  for aa in self.utils.added_keys]
-        self.assertTrue('r_corr' in self.utils.pysat_keys)
+        self.assertIn('r_corr', self.utils.pysat_keys)
         self.utils.pysat_keys[self.utils.pysat_keys.index("r_corr")] = None
 
         self.utils.test_ocb_added()
@@ -361,11 +372,12 @@ class TestPysatMethods(unittest.TestCase):
         """
         ocb_pysat.add_ocb_to_data(self.test_inst, "latitude", "mlt",
                                   evar_names=['vect_evar'],
-                                  vector_names={'vect_evar':
-                                                {'aacgm_n': 'dummy1',
-                                                 'aacgm_e': 'dummy2',
-                                                 'dat_name': 'vect',
-                                                 'dat_units': 'm/s'}},
+                                  vector_names={
+                                      'vect_evar':
+                                      {'aacgm_n': self.utils.pysat_key,
+                                       'aacgm_e': self.pysat_var2,
+                                       'dat_name': 'vect',
+                                       'dat_units': 'm/s'}},
                                   ocb=self.ocb)
 
         self.utils.added_keys = [kk for kk in self.test_inst.meta.keys()
@@ -378,8 +390,8 @@ class TestPysatMethods(unittest.TestCase):
             self.utils.arevectors.append(False)
             if pp == "r_corr":
                 pp = None
-            elif pp not in self.test_inst.data.columns:
-                pp = 'dummy1'
+            elif pp not in self.test_inst.variables:
+                pp = self.utils.pysat_key
                 self.utils.arevectors[-1] = True
             self.utils.pysat_keys.append(pp)
 
@@ -390,11 +402,12 @@ class TestPysatMethods(unittest.TestCase):
         """
         ocb_pysat.add_ocb_to_data(self.test_inst, "latitude", "mlt",
                                   curl_evar_names=['vect_cevar'],
-                                  vector_names={'vect_cevar':
-                                                {'aacgm_n': 'dummy1',
-                                                 'aacgm_e': 'dummy2',
-                                                 'dat_name': 'vect',
-                                                 'dat_units': 'm/s'}},
+                                  vector_names={
+                                      'vect_cevar':
+                                      {'aacgm_n': self.utils.pysat_key,
+                                       'aacgm_e': self.pysat_var2,
+                                       'dat_name': 'vect',
+                                       'dat_units': 'm/s'}},
                                   ocb=self.ocb)
 
         self.utils.added_keys = [kk for kk in self.test_inst.meta.keys()
@@ -407,8 +420,8 @@ class TestPysatMethods(unittest.TestCase):
             self.utils.arevectors.append(False)
             if pp == "r_corr":
                 pp = None
-            elif pp not in self.test_inst.data.columns:
-                pp = 'dummy1'
+            elif pp not in self.test_inst.variables:
+                pp = self.utils.pysat_key
                 self.utils.arevectors[-1] = True
             self.utils.pysat_keys.append(pp)
 
@@ -419,12 +432,13 @@ class TestPysatMethods(unittest.TestCase):
         """
 
         ocb_pysat.add_ocb_to_data(self.test_inst, "latitude", "mlt",
-                                  vector_names={'vect_cust':
-                                                {'aacgm_n': 'dummy1',
-                                                 'aacgm_e': 'dummy2',
-                                                 'dat_name': 'vect',
-                                                 'dat_units': 'm/s',
-                                                 'scale_func': None}},
+                                  vector_names={
+                                      'vect_cust':
+                                      {'aacgm_n': self.utils.pysat_key,
+                                       'aacgm_e': self.pysat_var2,
+                                       'dat_name': 'vect',
+                                       'dat_units': 'm/s',
+                                       'scale_func': None}},
                                   ocb=self.ocb)
 
         self.utils.added_keys = [kk for kk in self.test_inst.meta.keys()
@@ -437,8 +451,8 @@ class TestPysatMethods(unittest.TestCase):
             self.utils.arevectors.append(False)
             if pp == "r_corr":
                 pp = None
-            elif pp not in self.test_inst.data.columns:
-                pp = 'dummy1'
+            elif pp not in self.test_inst.variables:
+                pp = self.utils.pysat_key
                 self.utils.arevectors[-1] = True
             self.utils.pysat_keys.append(pp)
         self.utils.test_ocb_added()
@@ -447,14 +461,15 @@ class TestPysatMethods(unittest.TestCase):
         """ Test adding ocb to pysat with E-field, Curl, and Vector data
         """
         ocb_pysat.add_ocb_to_data(self.test_inst, "latitude", "mlt",
-                                  evar_names=['dummy1'],
-                                  curl_evar_names=['dummy2'],
-                                  vector_names={'vect_cust':
-                                                {'aacgm_n': 'dummy1',
-                                                 'aacgm_e': 'dummy2',
-                                                 'dat_name': 'vect',
-                                                 'dat_units': 'm/s',
-                                                 'scale_func': None}},
+                                  evar_names=[self.utils.pysat_key],
+                                  curl_evar_names=[self.pysat_var2],
+                                  vector_names={
+                                      'vect_cust':
+                                      {'aacgm_n': self.utils.pysat_key,
+                                       'aacgm_e': self.pysat_var2,
+                                       'dat_name': 'vect',
+                                       'dat_units': 'm/s',
+                                       'scale_func': None}},
                                   ocb=self.ocb)
 
         self.utils.added_keys = [kk for kk in self.test_inst.meta.keys()
@@ -467,8 +482,8 @@ class TestPysatMethods(unittest.TestCase):
             self.utils.arevectors.append(False)
             if pp == "r_corr":
                 pp = None
-            elif pp not in self.test_inst.data.columns:
-                pp = 'dummy1'
+            elif pp not in self.test_inst.variables:
+                pp = self.utils.pysat_key
                 self.utils.arevectors[-1] = True
             self.utils.pysat_keys.append(pp)
 
@@ -537,12 +552,97 @@ class TestPysatMethods(unittest.TestCase):
                                     'unknown vector name bad_n'):
             ocb_pysat.add_ocb_to_data(self.test_inst, "latitude", "mlt",
                                       evar_names=['bad'],
-                                      vector_names={'bad':
-                                                    {'aacgm_n': 'bad_n',
-                                                     'aacgm_e': 'dummy1',
-                                                     'dat_name': 'bad',
-                                                     'dat_units': ''}},
+                                      vector_names={
+                                          'bad':
+                                          {'aacgm_n': 'bad_n',
+                                           'aacgm_e': self.utils.pysat_key,
+                                           'dat_name': 'bad',
+                                           'dat_units': ''}},
                                       ocb=self.ocb)
+
+
+@unittest.skipIf(no_pysat, "pysat not installed")
+class TestPysatMethods2DXarray(TestPysatMethods):
+
+    def setUp(self):
+        """ Initialize the test class
+        """
+        self.test_file = path.join(path.dirname(ocbpy.__file__), "tests",
+                                   "test_data", "test_north_circle")
+        self.assertTrue(path.isfile(self.test_file))
+        self.ocb_kw = {"ocbfile": self.test_file,
+                       "instrument": "image", "hemisphere": 1}
+        self.ocb = ocbpy.OCBoundary(self.test_file, instrument='image',
+                                    hemisphere=1)
+        self.ocb.rec_ind = 27
+
+        self.test_inst = pysat.Instrument('pysat', 'testing2d_xarray',
+                                          num_samples=50400,
+                                          clean_level='clean',
+                                          update_files=True,
+                                          file_date_range=pds.date_range(
+                                              self.ocb.dtime[0],
+                                              self.ocb.dtime[-1], freq='1D'))
+        self.test_inst.load(date=self.ocb.dtime[self.ocb.rec_ind])
+        self.pysat_var2 = 'dummy2'
+
+        self.lwarn = u""
+        self.lout = u""
+        self.log_capture = StringIO()
+        ocbpy.logger.addHandler(logging.StreamHandler(self.log_capture))
+        ocbpy.logger.setLevel(logging.WARNING)
+
+        self.utils = TestPysatUtils("test_ocb_metadata")
+        self.utils.setUp()
+
+    def tearDown(self):
+        """ Tear down after each test
+        """
+        self.utils.tearDown()
+        del self.test_file, self.log_capture, self.ocb, self.test_inst
+        del self.utils, self.lout, self.lwarn, self.ocb_kw, self.pysat_var2
+
+
+@unittest.skipIf(no_pysat, "pysat not installed")
+class TestPysatMethodsXarray(TestPysatMethods):
+
+    def setUp(self):
+        """ Initialize the test class
+        """
+        self.test_file = path.join(path.dirname(ocbpy.__file__), "tests",
+                                   "test_data", "test_north_circle")
+        self.assertTrue(path.isfile(self.test_file))
+        self.ocb_kw = {"ocbfile": self.test_file,
+                       "instrument": "image", "hemisphere": 1}
+        self.ocb = ocbpy.OCBoundary(self.test_file, instrument='image',
+                                    hemisphere=1)
+        self.ocb.rec_ind = 27
+
+        self.test_inst = pysat.Instrument('pysat', 'testing_xarray',
+                                          num_samples=50400,
+                                          clean_level='clean',
+                                          update_files=True,
+                                          file_date_range=pds.date_range(
+                                              self.ocb.dtime[0],
+                                              self.ocb.dtime[-1], freq='1D'))
+        self.test_inst.load(date=self.ocb.dtime[self.ocb.rec_ind])
+        self.pysat_var2 = 'dummy2'
+
+        self.lwarn = u""
+        self.lout = u""
+        self.log_capture = StringIO()
+        ocbpy.logger.addHandler(logging.StreamHandler(self.log_capture))
+        ocbpy.logger.setLevel(logging.WARNING)
+
+        self.utils = TestPysatUtils("test_ocb_metadata")
+        self.utils.setUp()
+
+    def tearDown(self):
+        """ Tear down after each test
+        """
+        self.utils.tearDown()
+        del self.test_file, self.log_capture, self.ocb, self.test_inst
+        del self.utils, self.lout, self.lwarn, self.ocb_kw, self.pysat_var2
 
 
 @unittest.skipIf(no_pysat, "pysat not installed")
@@ -558,13 +658,13 @@ class TestPysatCustMethods(unittest.TestCase):
                                     hemisphere=1)
         self.ocb.rec_ind = 27
 
-        self.test_inst = pysat.Instrument(str('pysat'), str('testing'),
-                                          tag=str('50400'),
+        self.test_inst = pysat.Instrument('pysat', 'testing', num_samples=50400,
                                           clean_level='clean',
                                           update_files=True,
                                           file_date_range=pds.date_range(
                                               self.ocb.dtime[0],
                                               self.ocb.dtime[-1], freq='1D'))
+        self.pysat_var2 = 'dummy2'
         self.cust_kwargs = {'mlat_name': 'latitude', 'mlt_name': 'mlt',
                             'ocb': self.ocb}
 
@@ -580,20 +680,20 @@ class TestPysatCustMethods(unittest.TestCase):
     def tearDown(self):
         self.utils.tearDown()
         del self.test_file, self.log_capture, self.test_inst, self.ocb
-        del self.utils, self.lwarn, self.lout, self.cust_kwargs
+        del self.utils, self.lwarn, self.lout, self.cust_kwargs, self.pysat_var2
 
     def test_load(self):
         """ Test the pysat file loading"""
         self.test_inst.load(date=self.ocb.dtime[self.ocb.rec_ind])
         self.assertFalse(self.test_inst.empty)
-        self.assertTrue('latitude' in self.test_inst.data.columns)
-        self.assertTrue('mlt' in self.test_inst.data.columns)
+        self.assertIn('latitude', self.test_inst.variables)
+        self.assertIn('mlt', self.test_inst.variables)
 
     def test_cust_add_ocb_to_data_ocb_obj(self):
         """ Test adding ocb to pysat data with load using the loaded OCB object
         """
-        self.test_inst.custom.attach(ocb_pysat.add_ocb_to_data, 'modify',
-                                     **self.cust_kwargs)
+        self.test_inst.custom_attach(ocb_pysat.add_ocb_to_data,
+                                     kwargs=self.cust_kwargs)
 
         self.test_load()
 
@@ -602,7 +702,7 @@ class TestPysatCustMethods(unittest.TestCase):
         self.utils.nkeys = len(self.utils.added_keys)
         self.utils.pysat_keys = [aa.split("_ocb")[0]
                                  for aa in self.utils.added_keys]
-        self.assertTrue('r_corr' in self.utils.pysat_keys)
+        self.assertIn('r_corr', self.utils.pysat_keys)
         self.utils.pysat_keys[self.utils.pysat_keys.index("r_corr")] = None
 
         self.utils.test_inst = self.test_inst
@@ -617,8 +717,8 @@ class TestPysatCustMethods(unittest.TestCase):
         self.cust_kwargs['ocbfile'] = self.test_file
         self.cust_kwargs['instrument'] = 'image'
         self.cust_kwargs['hemisphere'] = 1
-        self.test_inst.custom.attach(ocb_pysat.add_ocb_to_data, 'modify',
-                                     **self.cust_kwargs)
+        self.test_inst.custom_attach(ocb_pysat.add_ocb_to_data,
+                                     kwargs=self.cust_kwargs)
 
         self.test_load()
 
@@ -627,7 +727,7 @@ class TestPysatCustMethods(unittest.TestCase):
         self.utils.nkeys = len(self.utils.added_keys)
         self.utils.pysat_keys = [aa.split("_ocb")[0]
                                  for aa in self.utils.added_keys]
-        self.assertTrue('r_corr' in self.utils.pysat_keys)
+        self.assertIn('r_corr', self.utils.pysat_keys)
         self.utils.pysat_keys[self.utils.pysat_keys.index("r_corr")] = None
         self.arevectors = [False for i in range(self.utils.nkeys)]
         self.utils.test_ocb_added()
@@ -635,9 +735,9 @@ class TestPysatCustMethods(unittest.TestCase):
     def test_cust_add_ocb_to_data_evar(self):
         """ Test adding ocb to pysat with load including E-field variables
         """
-        self.cust_kwargs['evar_names'] = ['dummy1']
-        self.test_inst.custom.attach(ocb_pysat.add_ocb_to_data, 'modify',
-                                     **self.cust_kwargs)
+        self.cust_kwargs['evar_names'] = [self.utils.pysat_key]
+        self.test_inst.custom_attach(ocb_pysat.add_ocb_to_data,
+                                     kwargs=self.cust_kwargs)
 
         self.test_load()
 
@@ -646,7 +746,7 @@ class TestPysatCustMethods(unittest.TestCase):
         self.utils.nkeys = len(self.utils.added_keys)
         self.utils.pysat_keys = [aa.split("_ocb")[0]
                                  for aa in self.utils.added_keys]
-        self.assertTrue('r_corr' in self.utils.pysat_keys)
+        self.assertIn('r_corr', self.utils.pysat_keys)
         self.utils.pysat_keys[self.utils.pysat_keys.index("r_corr")] = None
         self.arevectors = [False for i in range(self.utils.nkeys)]
 
@@ -656,9 +756,9 @@ class TestPysatCustMethods(unittest.TestCase):
         """ Test adding ocb to pysat with load including Curl E-field variables
         """
 
-        self.cust_kwargs['curl_evar_names'] = ['dummy2']
-        self.test_inst.custom.attach(ocb_pysat.add_ocb_to_data, 'modify',
-                                     **self.cust_kwargs)
+        self.cust_kwargs['curl_evar_names'] = [self.pysat_var2]
+        self.test_inst.custom_attach(ocb_pysat.add_ocb_to_data,
+                                     kwargs=self.cust_kwargs)
 
         self.test_load()
 
@@ -667,7 +767,7 @@ class TestPysatCustMethods(unittest.TestCase):
         self.utils.nkeys = len(self.utils.added_keys)
         self.utils.pysat_keys = [aa.split("_ocb")[0]
                                  for aa in self.utils.added_keys]
-        self.assertTrue('r_corr' in self.utils.pysat_keys)
+        self.assertIn('r_corr', self.utils.pysat_keys)
         self.utils.pysat_keys[self.utils.pysat_keys.index("r_corr")] = None
 
         self.utils.test_ocb_added()
@@ -678,12 +778,12 @@ class TestPysatCustMethods(unittest.TestCase):
 
         self.cust_kwargs['evar_names'] = ['vect_evar']
         self.cust_kwargs['vector_names'] = {'vect_evar':
-                                            {'aacgm_n': 'dummy1',
-                                             'aacgm_e': 'dummy2',
+                                            {'aacgm_n': self.utils.pysat_key,
+                                             'aacgm_e': self.pysat_var2,
                                              'dat_name': 'vect',
                                              'dat_units': 'm/s'}}
-        self.test_inst.custom.attach(ocb_pysat.add_ocb_to_data, 'modify',
-                                     **self.cust_kwargs)
+        self.test_inst.custom_attach(ocb_pysat.add_ocb_to_data,
+                                     kwargs=self.cust_kwargs)
 
         self.test_load()
 
@@ -697,8 +797,8 @@ class TestPysatCustMethods(unittest.TestCase):
             self.utils.arevectors.append(False)
             if pp == "r_corr":
                 pp = None
-            elif pp not in self.test_inst.data.columns:
-                pp = 'dummy1'
+            elif pp not in self.test_inst.variables:
+                pp = self.utils.pysat_key
                 self.utils.arevectors[-1] = True
             self.utils.pysat_keys.append(pp)
         self.utils.test_ocb_added()
@@ -708,12 +808,12 @@ class TestPysatCustMethods(unittest.TestCase):
         """
         self.cust_kwargs['curl_evar_names'] = ['vect_cevar']
         self.cust_kwargs['vector_names'] = {'vect_cevar':
-                                            {'aacgm_n': 'dummy1',
-                                             'aacgm_e': 'dummy2',
+                                            {'aacgm_n': self.utils.pysat_key,
+                                             'aacgm_e': self.pysat_var2,
                                              'dat_name': 'vect',
                                              'dat_units': 'm/s'}}
-        self.test_inst.custom.attach(ocb_pysat.add_ocb_to_data, 'modify',
-                                     **self.cust_kwargs)
+        self.test_inst.custom_attach(ocb_pysat.add_ocb_to_data,
+                                     kwargs=self.cust_kwargs)
 
         self.test_load()
 
@@ -727,8 +827,8 @@ class TestPysatCustMethods(unittest.TestCase):
             self.utils.arevectors.append(False)
             if pp == "r_corr":
                 pp = None
-            elif pp not in self.test_inst.data.columns:
-                pp = 'dummy1'
+            elif pp not in self.test_inst.variables:
+                pp = self.utils.pysat_key
                 self.utils.arevectors[-1] = True
             self.utils.pysat_keys.append(pp)
         self.utils.test_ocb_added()
@@ -737,13 +837,13 @@ class TestPysatCustMethods(unittest.TestCase):
         """ Test adding ocb to pysat with load including custom scaled variables
         """
         self.cust_kwargs['vector_names'] = {'vect_cust':
-                                            {'aacgm_n': 'dummy1',
-                                             'aacgm_e': 'dummy2',
+                                            {'aacgm_n': self.utils.pysat_key,
+                                             'aacgm_e': self.pysat_var2,
                                              'dat_name': 'vect',
                                              'dat_units': 'm/s',
                                              'scale_func': None}}
-        self.test_inst.custom.attach(ocb_pysat.add_ocb_to_data, 'modify',
-                                     **self.cust_kwargs)
+        self.test_inst.custom_attach(ocb_pysat.add_ocb_to_data,
+                                     kwargs=self.cust_kwargs)
         self.test_load()
 
         self.utils.added_keys = [kk for kk in self.test_inst.meta.keys()
@@ -756,8 +856,8 @@ class TestPysatCustMethods(unittest.TestCase):
             self.utils.arevectors.append(False)
             if pp == "r_corr":
                 pp = None
-            elif pp not in self.test_inst.data.columns:
-                pp = 'dummy1'
+            elif pp not in self.test_inst.variables:
+                pp = self.utils.pysat_key
                 self.utils.arevectors[-1] = True
             self.utils.pysat_keys.append(pp)
 
@@ -766,16 +866,16 @@ class TestPysatCustMethods(unittest.TestCase):
     def test_cust_add_ocb_to_data_all_types(self):
         """ Test adding ocb to pysat with load including E-field, Curl, & vects
         """
-        self.cust_kwargs['evar_names'] = ['dummy1']
-        self.cust_kwargs['curl_evar_names'] = ['dummy2']
+        self.cust_kwargs['evar_names'] = [self.utils.pysat_key]
+        self.cust_kwargs['curl_evar_names'] = [self.pysat_var2]
         self.cust_kwargs['vector_names'] = {'vect_cust':
-                                            {'aacgm_n': 'dummy1',
-                                             'aacgm_e': 'dummy2',
+                                            {'aacgm_n': self.utils.pysat_key,
+                                             'aacgm_e': self.pysat_var2,
                                              'dat_name': 'vect',
                                              'dat_units': 'm/s',
                                              'scale_func': None}}
-        self.test_inst.custom.attach(ocb_pysat.add_ocb_to_data, 'modify',
-                                     **self.cust_kwargs)
+        self.test_inst.custom_attach(ocb_pysat.add_ocb_to_data,
+                                     kwargs=self.cust_kwargs)
 
         self.test_load()
 
@@ -789,8 +889,8 @@ class TestPysatCustMethods(unittest.TestCase):
             self.utils.arevectors.append(False)
             if pp == "r_corr":
                 pp = None
-            elif pp not in self.test_inst.data.columns:
-                pp = 'dummy1'
+            elif pp not in self.test_inst.variables:
+                pp = self.utils.pysat_key
                 self.utils.arevectors[-1] = True
             self.utils.pysat_keys.append(pp)
 
@@ -803,8 +903,8 @@ class TestPysatCustMethods(unittest.TestCase):
         self.cust_kwargs['ocbfile'] = None
         self.cust_kwargs['instrument'] = 'image'
         self.cust_kwargs['hemisphere'] = 1
-        self.test_inst.custom.attach(ocb_pysat.add_ocb_to_data, 'modify',
-                                     **self.cust_kwargs)
+        self.test_inst.custom_attach(ocb_pysat.add_ocb_to_data,
+                                     kwargs=self.cust_kwargs)
 
         self.test_load()
 
@@ -825,8 +925,8 @@ class TestPysatCustMethods(unittest.TestCase):
         """ Test failure of unknown mlt key in add_ocb_to_data in custom func
         """
         self.cust_kwargs['mlt_name'] = 'bad'
-        self.test_inst.custom.attach(ocb_pysat.add_ocb_to_data, 'modify',
-                                     **self.cust_kwargs)
+        self.test_inst.custom_attach(ocb_pysat.add_ocb_to_data,
+                                     kwargs=self.cust_kwargs)
 
         with self.assertRaisesRegex(ValueError,
                                     'unknown magnetic local time name bad'):
@@ -836,8 +936,8 @@ class TestPysatCustMethods(unittest.TestCase):
         """ Test failure of unknown E field key in custom func
         """
         self.cust_kwargs['evar_names'] = ['bad']
-        self.test_inst.custom.attach(ocb_pysat.add_ocb_to_data, 'modify',
-                                     **self.cust_kwargs)
+        self.test_inst.custom_attach(ocb_pysat.add_ocb_to_data,
+                                     kwargs=self.cust_kwargs)
 
         with self.assertRaisesRegex(ValueError,
                                     'at least one unknown E field name'):
@@ -847,8 +947,8 @@ class TestPysatCustMethods(unittest.TestCase):
         """ Test failure of unknown E field key in custom func
         """
         self.cust_kwargs['curl_evar_names'] = ['bad']
-        self.test_inst.custom.attach(ocb_pysat.add_ocb_to_data, 'modify',
-                                     **self.cust_kwargs)
+        self.test_inst.custom_attach(ocb_pysat.add_ocb_to_data,
+                                     kwargs=self.cust_kwargs)
 
         with self.assertRaisesRegex(ValueError,
                                     'at least one unknown E field name'):
@@ -861,8 +961,8 @@ class TestPysatCustMethods(unittest.TestCase):
                                                     'aacgm_e': 'bad_e',
                                                     'dat_name': 'bad',
                                                     'dat_units': ''}}
-        self.test_inst.custom.attach(ocb_pysat.add_ocb_to_data, 'modify',
-                                     **self.cust_kwargs)
+        self.test_inst.custom_attach(ocb_pysat.add_ocb_to_data,
+                                     kwargs=self.cust_kwargs)
 
         with self.assertRaisesRegex(ValueError,
                                     'missing scaling function for bad'):
@@ -872,12 +972,91 @@ class TestPysatCustMethods(unittest.TestCase):
         """ Test failure of missing scaling function in custom func
         """
         self.cust_kwargs['evar_names'] = ['bad']
-        self.cust_kwargs['vector_names'] = {'bad': {'aacgm_n': 'bad_n',
-                                                    'aacgm_e': 'dummy1',
-                                                    'dat_name': 'bad',
-                                                    'dat_units': ''}}
-        self.test_inst.custom.attach(ocb_pysat.add_ocb_to_data, 'modify',
-                                     **self.cust_kwargs)
+        self.cust_kwargs['vector_names'] = {'bad':
+                                            {'aacgm_n': 'bad_n',
+                                             'aacgm_e': self.utils.pysat_key,
+                                             'dat_name': 'bad',
+                                             'dat_units': ''}}
+        self.test_inst.custom_attach(ocb_pysat.add_ocb_to_data,
+                                     kwargs=self.cust_kwargs)
 
         with self.assertRaisesRegex(ValueError, 'unknown vector name bad_n'):
             self.test_load()
+
+
+@unittest.skipIf(no_pysat, "pysat not installed")
+class TestPysatCustMethodsXarray(TestPysatCustMethods):
+
+    def setUp(self):
+        """ Initialize the unit tests for using the pysat.Custom methods
+        """
+        self.test_file = path.join(path.dirname(ocbpy.__file__), "tests",
+                                   "test_data", "test_north_circle")
+        self.assertTrue(path.isfile(self.test_file))
+        self.ocb = ocbpy.OCBoundary(self.test_file, instrument='image',
+                                    hemisphere=1)
+        self.ocb.rec_ind = 27
+
+        self.test_inst = pysat.Instrument('pysat', 'testing_xarray',
+                                          num_samples=50400,
+                                          clean_level='clean',
+                                          update_files=True,
+                                          file_date_range=pds.date_range(
+                                              self.ocb.dtime[0],
+                                              self.ocb.dtime[-1], freq='1D'))
+        self.pysat_var2 = 'dummy2'
+        self.cust_kwargs = {'mlat_name': 'latitude', 'mlt_name': 'mlt',
+                            'ocb': self.ocb}
+
+        self.lwarn = u""
+        self.lout = u""
+        self.log_capture = StringIO()
+        ocbpy.logger.addHandler(logging.StreamHandler(self.log_capture))
+        ocbpy.logger.setLevel(logging.WARNING)
+
+        self.utils = TestPysatUtils("test_ocb_metadata")
+        self.utils.setUp()
+
+    def tearDown(self):
+        self.utils.tearDown()
+        del self.test_file, self.log_capture, self.test_inst, self.ocb
+        del self.utils, self.lwarn, self.lout, self.cust_kwargs, self.pysat_var2
+
+
+@unittest.skipIf(no_pysat, "pysat not installed")
+class TestPysatCustMethods2DXarray(TestPysatCustMethods):
+
+    def setUp(self):
+        """ Initialize the unit tests for using the pysat.Custom methods
+        """
+        self.test_file = path.join(path.dirname(ocbpy.__file__), "tests",
+                                   "test_data", "test_north_circle")
+        self.assertTrue(path.isfile(self.test_file))
+        self.ocb = ocbpy.OCBoundary(self.test_file, instrument='image',
+                                    hemisphere=1)
+        self.ocb.rec_ind = 27
+
+        self.test_inst = pysat.Instrument('pysat', 'testing2d_xarray',
+                                          num_samples=50400,
+                                          clean_level='clean',
+                                          update_files=True,
+                                          file_date_range=pds.date_range(
+                                              self.ocb.dtime[0],
+                                              self.ocb.dtime[-1], freq='1D'))
+        self.pysat_var2 = 'dummy2'
+        self.cust_kwargs = {'mlat_name': 'latitude', 'mlt_name': 'mlt',
+                            'ocb': self.ocb}
+
+        self.lwarn = u""
+        self.lout = u""
+        self.log_capture = StringIO()
+        ocbpy.logger.addHandler(logging.StreamHandler(self.log_capture))
+        ocbpy.logger.setLevel(logging.WARNING)
+
+        self.utils = TestPysatUtils("test_ocb_metadata")
+        self.utils.setUp()
+
+    def tearDown(self):
+        self.utils.tearDown()
+        del self.test_file, self.log_capture, self.test_inst, self.ocb
+        del self.utils, self.lwarn, self.lout, self.cust_kwargs, self.pysat_var2